Kill attachListener_aix.cpp
